requirements.txt
# Core Testing Framework
pytest==7.4.4
pytest-html==4.1.1
pytest-metadata==3.1.1

# Allure Reporting
allure-pytest==2.15.0

# Test Ordering
pytest-ordering==0.6

# HTTP Client
requests==2.31.0

# Utility Libraries
urllib3==2.1.0

# Optional: For parallel test execution
# pytest-xdist==3.5.0

# Optional: For test coverage
# pytest-cov==4.1.0
